pyproject.toml
[tool.poetry]
name = "qa-from-pdf-file"
version = "0.1.0"
description = ""
authors = ["Christian Laurean <christianlaurean1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
langchain = "^0.2.15"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.15"
python-dotenv = "^1.0.1"
langchain-core = "^0.2.37"
langchain-text-splitters = "^0.2.2"
pypdf = "^4.3.1"
langchain-chroma = "^0.1.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
